requirements.txt
pysqlite3
openpyxl
extra-streamlit-components
bcrypt
PyJWT
PyYAML